Rename faucet to mock for tests
